requirements.txt
deepcell==0.12.9
google-cloud-bigquery
google-cloud-notebooks
gs-fastcopy
imagecodecs
JSON-log-formatter
numpy
protobuf==3.20.3
pydantic
pytest
requests
smart_open[gcs]
snakeviz
tenacity
tifffile
